BigQuery_jsondata/requirements.txt
google-cloud-bigquery
google-auth
requests
